Refactor integration tests to use new testing module
