+++ src/pynwb/base.py
@@ -155,9 +155,9 @@ class TimeSeries(NWBDataInterface):
             self.rate = rate
             if starting_time is not None:
                 self.starting_time = starting_time
-                self.starting_time_unit = self.__time_unit
             else:
                 self.starting_time = 0.0
+            self.starting_time_unit = self.__time_unit
         else:
             raise TypeError("either 'timestamps' or 'rate' must be specified")
