+++ src/pynwb/testing/__init__.py
@@ -1,5 +1,5 @@
-from hdmf.testing import TestCase, TestMapH5RoundTrip
-from .testmap import TestMapNWBContainer, TestMapRoundTrip, TestDataInterfaceIO
+from hdmf.testing import TestCase, TestH5RoundTripMixin
+from .testh5io import TestNWBH5IOMixin, TestAcquisitionH5IOMixin
 from .utils import remove_test_file
 
 CORE_NAMESPACE = 'core'
