+++ tests/integration/hdf5/test_ecephys.py
@@ -1,67 +1,64 @@
-import unittest
-
 from hdmf.common import DynamicTableRegion
 
 from pynwb.ecephys import ElectrodeGroup, ElectricalSeries, FilteredEphys, LFP, Clustering, ClusterWaveforms,\
                           SpikeEventSeries, EventWaveform, EventDetection, FeatureExtraction
 from pynwb.device import Device
 from pynwb.file import ElectrodeTable as get_electrode_table
-from pynwb.testing import TestMapRoundTrip, TestDataInterfaceIO
-
-from abc import ABCMeta
+from pynwb.testing import TestNWBH5IOMixin, TestAcquisitionH5IOMixin, TestCase
 
 
-class TestElectrodeGroupIO(TestMapRoundTrip):
+class TestElectrodeGroupIO(TestNWBH5IOMixin, TestCase):
 
     def setUpContainer(self):
+        """ Return the test ElectrodeGroup to read/write """
         self.dev1 = Device('dev1')
-        return ElectrodeGroup('elec1', 'a test ElectrodeGroup',
-                                       'a nonexistent place',
-                                       self.dev1)
+        eg = ElectrodeGroup('elec1', 'a test ElectrodeGroup', 'a nonexistent place', self.dev1)
+        return eg
 
     def addContainer(self, nwbfile):
-        ''' Should take an NWBFile object and add the container to it '''
+        """ Add the test ElectrodeGroup to the given NWBFile """
         nwbfile.add_device(self.dev1)
         nwbfile.add_electrode_group(self.container)
 
     def getContainer(self, nwbfile):
-        ''' Should take an NWBFile object and return the Container'''
+        """ Return the test ElectrodeGroup from the given NWBFile """
         return nwbfile.get_electrode_group(self.container.name)
 
 
-class TestElectricalSeriesIO(TestDataInterfaceIO):
+class TestElectricalSeriesIO(TestAcquisitionH5IOMixin, TestCase):
 
     @staticmethod
     def make_electrode_table(self):
+        """ Make an electrode table, electrode group, and device """
         self.table = get_electrode_table()
         self.dev1 = Device('dev1')
         self.group = ElectrodeGroup('tetrode1',
                                     'tetrode description', 'tetrode location', self.dev1)
-        self.table.add_row(id=1, x=1.0, y=2.0, z=3.0, imp=-1.0, location='CA1', filtering='none', group=self.group,
-                           group_name='tetrode1')
-        self.table.add_row(id=2, x=1.0, y=2.0, z=3.0, imp=-2.0, location='CA1', filtering='none', group=self.group,
-                           group_name='tetrode1')
-        self.table.add_row(id=3, x=1.0, y=2.0, z=3.0, imp=-3.0, location='CA1', filtering='none', group=self.group,
-                           group_name='tetrode1')
-        self.table.add_row(id=4, x=1.0, y=2.0, z=3.0, imp=-4.0, location='CA1', filtering='none', group=self.group,
-                           group_name='tetrode1')
+        for i in range(4):
+            self.table.add_row(x=i, y=2.0, z=3.0, imp=-1.0, location='CA1', filtering='none', group=self.group,
+                               group_name='tetrode1')
 
     def setUpContainer(self):
+        """ Return the test ElectricalSeries to read/write """
         self.make_electrode_table(self)
         region = DynamicTableRegion('electrodes', [0, 2], 'the first and third electrodes', self.table)
         data = list(zip(range(10), range(10, 20)))
         timestamps = list(map(lambda x: x/10, range(10)))
-        ret = ElectricalSeries('test_eS', data, region, channel_conversion=[4., .4], timestamps=timestamps)
-        return ret
+        es = ElectricalSeries('test_eS', data, region, channel_conversion=[4., .4], timestamps=timestamps)
+        return es
 
     def addContainer(self, nwbfile):
-        ''' Should take an NWBFile object and add the container to it '''
+        """ Add the test ElectricalSeries and related objects to the given NWBFile """
         nwbfile.add_device(self.dev1)
         nwbfile.add_electrode_group(self.group)
         nwbfile.set_electrode_table(self.table)
         nwbfile.add_acquisition(self.container)
 
     def test_eg_ref(self):
+        """
+        Test that the electrode DynamicTableRegion references of the read ElectricalSeries have a group that
+        correctly resolves to ElectrodeGroup instances.
+        """
         read = self.roundtripContainer()
         row1 = read.electrodes[0]
         row2 = read.electrodes[1]
@@ -69,9 +66,17 @@ class TestElectricalSeriesIO(TestDataInterfaceIO):
         self.assertIsInstance(row2.iloc[0]['group'], ElectrodeGroup)
 
 
-class TestMultiElectricalSeries(TestDataInterfaceIO, metaclass=ABCMeta):
+class TestMultiElectricalSeriesIOMixin(TestAcquisitionH5IOMixin):
+    """
+    Mixin class for methods to run a roundtrip test writing an NWB file with multiple ElectricalSeries.
 
-    def setUpElectricalSeriesContainers(self):
+    The abstract method setUpContainer needs to be implemented by classes that include this mixin.
+        def setUpContainer(self):
+            # return a test Container to read/write
+    """
+
+    def setUpTwoElectricalSeries(self):
+        """ Return two test ElectricalSeries to read/write """
         TestElectricalSeriesIO.make_electrode_table(self)
         region1 = DynamicTableRegion('electrodes', [0, 2], 'the first and third electrodes', self.table)
         region2 = DynamicTableRegion('electrodes', [1, 3], 'the second and fourth electrodes', self.table)
@@ -83,64 +88,67 @@ class TestMultiElectricalSeries(TestDataInterfaceIO, metaclass=ABCMeta):
         return es1, es2
 
     def addContainer(self, nwbfile):
-        ''' Should take an NWBFile object and add the container to it '''
+        """ Add the test ElectricalSeries and related objects to the given NWBFile """
         nwbfile.add_device(self.dev1)
         nwbfile.add_electrode_group(self.group)
         nwbfile.set_electrode_table(self.table)
         nwbfile.add_acquisition(self.container)
 
-    def setUpContainer(self):
-        raise unittest.SkipTest('Cannot run test unless addContainer is implemented')
 
-
-class TestLFP(TestMultiElectricalSeries):
+class TestLFPIO(TestMultiElectricalSeriesIOMixin, TestCase):
 
     def setUpContainer(self):
-        es = self.setUpElectricalSeriesContainers()
-        ret = LFP(es)
-        return ret
+        """ Return a test LFP to read/write """
+        es = self.setUpTwoElectricalSeries()
+        lfp = LFP(es)
+        return lfp
 
 
-class TestFilteredEphys(TestMultiElectricalSeries):
-
-    def setUpContainer(self):
-        es = self.setUpElectricalSeriesContainers()
-        ret = FilteredEphys(es)
-        return ret
-
-
-class TestClusteringIO(TestDataInterfaceIO):
+class TestFilteredEphysIO(TestMultiElectricalSeriesIOMixin, TestCase):
 
     def setUpContainer(self):
+        """ Return a test FilteredEphys to read/write """
+        es = self.setUpTwoElectricalSeries()
+        fe = FilteredEphys(es)
+        return fe
+
+
+class TestClusteringIO(TestAcquisitionH5IOMixin, TestCase):
+
+    def setUpContainer(self):
+        """ Return a test Clustering to read/write """
         with self.assertWarnsWith(DeprecationWarning, 'use pynwb.misc.Units or NWBFile.units instead'):
             return Clustering("A fake Clustering interface",
                               [0, 1, 2, 0, 1, 2], [100., 101., 102.], [float(i) for i in range(10, 61, 10)])
 
     def roundtripContainer(self, cache_spec=False):
-        # self.reader.read() will throw DeprecationWarning when reading the Clustering object
+        # catch the DeprecationWarning raised when reading the Clustering object from file
         with self.assertWarnsWith(DeprecationWarning, 'use pynwb.misc.Units or NWBFile.units instead'):
-            return super(TestClusteringIO, self).roundtripContainer(cache_spec)
+            return super().roundtripContainer(cache_spec)
 
 
-class EventWaveformConstructor(TestDataInterfaceIO):
+class EventWaveformConstructor(TestAcquisitionH5IOMixin, TestCase):
+
     def setUpContainer(self):
+        """ Return a test EventWaveform to read/write """
         TestElectricalSeriesIO.make_electrode_table(self)
         region = DynamicTableRegion('electrodes', [0, 2], 'the first and third electrodes', self.table)
-        sES = SpikeEventSeries(
-            'test_sES', ((1, 1, 1), (2, 2, 2)), list(range(2)), region)
+        sES = SpikeEventSeries('test_sES', ((1, 1, 1), (2, 2, 2)), list(range(2)), region)
         ew = EventWaveform(sES)
         return ew
 
     def addContainer(self, nwbfile):
-        ''' Should take an NWBFile object and add the container to it '''
+        """ Add the test EventWaveform and related objects to the given NWBFile """
         nwbfile.add_device(self.dev1)
         nwbfile.add_electrode_group(self.group)
         nwbfile.set_electrode_table(self.table)
         nwbfile.add_acquisition(self.container)
 
 
-class ClusterWaveformsConstructor(TestDataInterfaceIO):
+class ClusterWaveformsConstructor(TestAcquisitionH5IOMixin, TestCase):
+
     def setUpContainer(self):
+        """ Return a test ClusterWaveforms to read/write """
         times = [1.3, 2.3]
         num = [3, 4]
         peak_over_rms = [5.3, 6.3]
@@ -153,18 +161,20 @@ class ClusterWaveformsConstructor(TestDataInterfaceIO):
         return cw
 
     def addContainer(self, nwbfile):
-        ''' Should take an NWBFile object and add the container to it '''
+        """ Add the test ClusterWaveforms and related objects to the given NWBFile """
         nwbfile.add_acquisition(self.clustering)
         nwbfile.add_acquisition(self.container)
 
     def roundtripContainer(self, cache_spec=False):
-        # self.reader.read() will throw DeprecationWarning when reading the ClusterWaveformsConstructor object
+        # catch the DeprecationWarning raised when reading the Clustering object from file
         with self.assertWarnsWith(DeprecationWarning, 'use pynwb.misc.Units or NWBFile.units instead'):
-            return super(ClusterWaveformsConstructor, self).roundtripContainer(cache_spec)
+            return super().roundtripContainer(cache_spec)
 
 
-class FeatureExtractionConstructor(TestDataInterfaceIO):
+class FeatureExtractionConstructor(TestAcquisitionH5IOMixin, TestCase):
+
     def setUpContainer(self):
+        """ Return a test FeatureExtraction to read/write """
         event_times = [1.9, 3.5]
         TestElectricalSeriesIO.make_electrode_table(self)
         region = DynamicTableRegion('electrodes', [0, 2], 'the first and third electrodes', self.table)
@@ -174,25 +184,27 @@ class FeatureExtractionConstructor(TestDataInterfaceIO):
         return fe
 
     def addContainer(self, nwbfile):
-        ''' Should take an NWBFile object and add the container to it '''
+        """ Add the test FeatureExtraction and related objects to the given NWBFile """
         nwbfile.add_device(self.dev1)
         nwbfile.add_electrode_group(self.group)
         nwbfile.set_electrode_table(self.table)
         nwbfile.add_acquisition(self.container)
 
 
-class EventDetectionConstructor(TestDataInterfaceIO):
+class EventDetectionConstructor(TestAcquisitionH5IOMixin, TestCase):
+
     def setUpContainer(self):
-        data = list(range(10))
-        ts = [0.1, 0.2, 0.3, 0.4, 0.5, 0.6, 0.7, 0.8, 0.9, 1.0]
+        """ Return a test EventDetection to read/write """
         TestElectricalSeriesIO.make_electrode_table(self)
         region = DynamicTableRegion('electrodes', [0, 2], 'the first and third electrodes', self.table)
+        data = list(range(10))
+        ts = [0.1, 0.2, 0.3, 0.4, 0.5, 0.6, 0.7, 0.8, 0.9, 1.0]
         self.eS = ElectricalSeries('test_eS', data, region, timestamps=ts)
         eD = EventDetection('detection_method', self.eS, (1, 2, 3), (0.1, 0.2, 0.3))
         return eD
 
     def addContainer(self, nwbfile):
-        ''' Should take an NWBFile object and add the container to it '''
+        """ Add the test EventDetection and related objects to the given NWBFile """
         nwbfile.add_device(self.dev1)
         nwbfile.add_electrode_group(self.group)
         nwbfile.set_electrode_table(self.table)
