+++ tests/integration/hdf5/test_image.py
@@ -1,12 +1,13 @@
 import numpy as np
 
 from pynwb.image import OpticalSeries
-from pynwb.testing import TestMapRoundTrip
+from pynwb.testing import TestNWBH5IOMixin, TestCase
 
 
-class TestOpticalSeries(TestMapRoundTrip):
+class TestOpticalSeriesIO(TestNWBH5IOMixin, TestCase):
 
     def setUpContainer(self):
+        """ Return a test OpticalSeries to read/write """
         self.optical_series = OpticalSeries(name='OpticalSeries',
                                             distance=8.,
                                             field_of_view=(4., 5.),
@@ -18,9 +19,9 @@ class TestOpticalSeries(TestMapRoundTrip):
         return self.optical_series
 
     def addContainer(self, nwbfile):
-        ''' Should take an NWBFile object and add the container to it '''
+        """ Add the test OpticalSeries to the given NWBFile """
         nwbfile.add_stimulus(self.optical_series)
 
     def getContainer(self, nwbfile):
-        ''' Should take an NWBFile object and return the Container'''
+        """ Return the test OpticalSeries from the given NWBFile """
         return nwbfile.stimulus['OpticalSeries']
